Remove helm values cycle
Adding .global when ranging over machineDeployments causes a cycle to
happen if using the default values. This seems to only happen when doing
a `helm upgrade` and we're still not sure why. Redefining the
`.Values.machineDeployments` var to anything fixes this issue, however
using the defaults should also work.
To fix this we must call all the functions in from helpers inside the
range, so we keep the global scope. If the functions are called inside
another function inside the range, then the global scope is lost.

Co-authored-by: Jose Armesto <jose@giantswarm.io>